Rename the action file to match the component name.
