+++ src/smart-components/request/request-detail/action-transcript.js
