+++ src/smart-components/request/request-detail/request-stage-transcript.js
@@ -4,9 +4,10 @@ import { Title } from '@patternfly/react-core';
 import StageList from './stage-list';
 
 const RequestStageTranscript = ({ request }) => {
+  console.log('Debug - request', request);
   return (<Fragment>
     <Title size="sm" style={ { paddingLeft: '32px' } }> Stage(s) transcript </Title>
-    <StageList items={ request ? request.stages : [] } noItems={ 'No Stages' } />
+    <StageList items={ request ? request.stages : [] } noItems={ 'No Stages' } active_stage={ request.active_stage }/>
   </Fragment>);
 };
